Module hydroflow::builder

Hydroflow’s Surface API.

Internal Documentation

Due to the limitations of type-level programming in Rust, this code is “baklava” code containing lot of layers. Each layer does one thing, then constructs the next layer(s) down. This table describes what each layer does and is named. Layers are listed starting from the highest (user-facing API) layer and ending with the lowest (code-running) layer.

Layer Structure

              (A) Surface API
        (B) (Push Surface Reversed*)
                     |
           (D) Subgraph Builder
                     |
          (E) Iterator/Pusherator

*Only used with Push to reverse the ownership direction.

(Note (C) used to be Connectors but they’re no longer needed)

Layer Descriptions

Layer 	Purpose 	Naming

Pull 	Push

(A) The Surface API 	 • Presents a clean functional-chaining API for users.
• Consumed to simultaneously create a (C) connector and (D) builder.
• Push Only: Extra layer needed to reverse ownership order. 	PullSurface 	PushSurface

(B) Push Surface Reversed 	N/A 	PushSurfaceReversed

(D) Subgraph Builders 	 • On each subgraph invocation, constructs the (E) iterators and pivot which will be run.
• Is owned by the subgraph task, lends closures to (E) iterators.
• Uses the input/output HandoffList variadic type. 	PullBuild 	PushBuild

(E) Iterators 	 • Runs code on individual dataflow elements, in the case of dataflow.
• In the future, will correspond to semilattice morphisms alternatively. 	std::iter::Iterator 	crate::compiled::Pusherator

How the code actually runs

The layers are used in HydroflowBuilder::add_subgraph. The method receives a pivot with PullSurface and PushSurfaceReversed halves. Then into_parts() splits them into the ports half and the build half. Those are then both sent to Hydroflow::add_subgraph() to create a subgraph.

Modules

build

Internal “subgraph builders” to implement the Surface API. For more info see super.

prelude

Prelude to quick-import items needed for using the Surface API.

surface

Structs used to create the user-facing Surface API.

Structs

HydroflowBuilder

The user-facing entry point for the Surface API.

Traits

IntoHydroflow

A trait to convert IntoIterators for use with Hydroflow’s surface (build) API.