Module hydroflow::builder::surface

Structs used to create the user-facing Surface API.

Main user-facing traits are BaseSurface, PullSurface, and PushSurface, which provide an iterator-like API with easy method chaining. The traits need to be imported when using the Surface API. You can use the prelude to do this easily:

use hydroflow::build::prelude::*;

BaseSurface provides linear chaining methods like BaseSurface::map, BaseSurface::filter, etc..

PullSurface provides methods to combine multiple input streams: PullSurface::chain, PullSurface::join.

To switch to push, call PullSurface::pull_to_push.

PushSurface provides sink chaining methods and methods to split into multiple output streams: PushSurface::tee, PushSurface::for_each.

For implementation info see super.

Modules

exchange

filter

filter_map

flatten

map

map_scan

pivot

pull_batch

pull_chain

pull_cross_join

pull_fold_epoch

pull_half_hash_join

pull_handoff

pull_iter

pull_join

push_for_each

push_handoff

push_partition

push_pivot

push_start

push_tee

Traits

AssembleFlowGraph

BaseSurface

Common trait shared between push and pull surface APIs.

PullSurface

PushSurface

PushSurfaceReversed

This extra layer is needed due to the ownership order. In the functional chaining syntax each operator owns the previous (can only go in order things are called/defined), but in the end we need each pusherator to own the next pusherator which it’s pushing to.

Type Definitions

FilterMapNoCtxFn

FilterNoCtxFn

ForEachNoCtxFunc

InspectMapFunc

InspectMapNoCtxFunc

MapNoCtxFn

PartitionNoCtxOutput