Trait hydroflow::builder::surface::PullSurface

pub trait PullSurface: BaseSurface {
    type InputHandoffs: PortList<RECV>;
    type Build: PullBuild<InputHandoffs = Self::InputHandoffs, ItemOut = Self::ItemOut>;

    fn make_parts(self, ctx: &mut Context) -> (Self::InputHandoffs, Self::Build);

    fn chain<Other>(self, other: Other) -> ChainPullSurface<Self, Other>
   where
        Self: Sized,
        Other: PullSurface<ItemOut = Self::ItemOut>,
        Self::InputHandoffs: Extend<Other::InputHandoffs>,
        <Self::InputHandoffs as Extend<Other::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Self::InputHandoffs, Suffix = Other::InputHandoffs>,
    { ... }
    fn join<Other, Key, ValSelf, ValOther>(
        self,
        other: Other
    ) -> JoinPullSurface<Self, Other>
   where
        Self: Sized + PullSurface<ItemOut = (Key, ValSelf)>,
        Other: PullSurface<ItemOut = (Key, ValOther)>,
        Key: 'static + Eq + Hash + Clone,
        ValSelf: 'static + Eq + Clone,
        ValOther: 'static + Eq + Clone,
        Self::InputHandoffs: Extend<Other::InputHandoffs>,
        <Self::InputHandoffs as Extend<Other::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Self::InputHandoffs, Suffix = Other::InputHandoffs>,
    { ... }
    fn batch_with<Other, L, Update, Tick>(
        self,
        other: Other
    ) -> BatchPullSurface<Self, Other, L, Update, Tick>
   where
        Self: Sized + PullSurface<ItemOut = Update::Repr>,
        Other: PullSurface<ItemOut = Tick>,
        Update: 'static + LatticeRepr,
        L: 'static + LatticeRepr + Merge<Update>,
        Self::InputHandoffs: Extend<Other::InputHandoffs>,
        <Self::InputHandoffs as Extend<Other::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Self::InputHandoffs, Suffix = Other::InputHandoffs>,
    { ... }
    fn half_hash_join<Other, Key, ValSelf, L, Update>(
        self,
        other: Other
    ) -> HalfHashJoinPullSurface<Self, Other, L, Update>
   where
        Self: Sized + PullSurface<ItemOut = (Key, ValSelf)>,
        Other: PullSurface<ItemOut = (Key, Update::Repr)>,
        Key: 'static + Eq + Hash,
        ValSelf: 'static,
        Update: 'static + LatticeRepr,
        L: 'static + LatticeRepr + Merge<Update>,
        L::Repr: Clone,
        Other::InputHandoffs: Extend<Self::InputHandoffs>,
        <Other::InputHandoffs as Extend<Self::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Other::InputHandoffs, Suffix = Self::InputHandoffs>,
    { ... }
    fn cross_join<Other>(self, other: Other) -> CrossJoinPullSurface<Self, Other>
   where
        Self: Sized,
        Other: PullSurface,
        Self::ItemOut: 'static + Eq + Clone,
        Other::ItemOut: 'static + Eq + Clone,
        Self::InputHandoffs: Extend<Other::InputHandoffs>,
        <Self::InputHandoffs as Extend<Other::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Self::InputHandoffs, Suffix = Other::InputHandoffs>,
    { ... }
    fn fold_epoch<Init, Func, Out>(
        self,
        init: Init,
        func: Func
    ) -> FoldEpochPullSurface<Self, Init, Func>
   where
        Self: Sized,
        Init: FnMut(&Context) -> Out,
        Func: FnMut(&Context, Out, Self::ItemOut) -> Out,
    { ... }
    fn pull_to_push(self) -> PivotPushSurface<Self>
   where
        Self: Sized,
    { ... }
}

Required Associated Types

type InputHandoffs: PortList<RECV>

type Build: PullBuild<InputHandoffs = Self::InputHandoffs, ItemOut = Self::ItemOut>

Required Methods

fn make_parts(self, ctx: &mut Context) -> (Self::InputHandoffs, Self::Build)

Provided Methods

fn chain<Other>(self, other: Other) -> ChainPullSurface<Self, Other>where
Self: Sized,
Other: PullSurface<ItemOut = Self::ItemOut>,
Self::InputHandoffs: Extend<Other::InputHandoffs>,
<Self::InputHandoffs as Extend<Other::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Self::InputHandoffs, Suffix = Other::InputHandoffs>,

fn join<Other, Key, ValSelf, ValOther>(
self,
other: Other
) -> JoinPullSurface<Self, Other>where
Self: Sized + PullSurface<ItemOut = (Key, ValSelf)>,
Other: PullSurface<ItemOut = (Key, ValOther)>,
Key: 'static + Eq + Hash + Clone,
ValSelf: 'static + Eq + Clone,
ValOther: 'static + Eq + Clone,
Self::InputHandoffs: Extend<Other::InputHandoffs>,
<Self::InputHandoffs as Extend<Other::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Self::InputHandoffs, Suffix = Other::InputHandoffs>,

fn batch_with<Other, L, Update, Tick>(
self,
other: Other
) -> BatchPullSurface<Self, Other, L, Update, Tick>where
Self: Sized + PullSurface<ItemOut = Update::Repr>,
Other: PullSurface<ItemOut = Tick>,
Update: 'static + LatticeRepr,
L: 'static + LatticeRepr + Merge<Update>,
Self::InputHandoffs: Extend<Other::InputHandoffs>,
<Self::InputHandoffs as Extend<Other::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Self::InputHandoffs, Suffix = Other::InputHandoffs>,

fn half_hash_join<Other, Key, ValSelf, L, Update>(
self,
other: Other
) -> HalfHashJoinPullSurface<Self, Other, L, Update>where
Self: Sized + PullSurface<ItemOut = (Key, ValSelf)>,
Other: PullSurface<ItemOut = (Key, Update::Repr)>,
Key: 'static + Eq + Hash,
ValSelf: 'static,
Update: 'static + LatticeRepr,
L: 'static + LatticeRepr + Merge<Update>,
L::Repr: Clone,
Other::InputHandoffs: Extend<Self::InputHandoffs>,
<Other::InputHandoffs as Extend<Self::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Other::InputHandoffs, Suffix = Self::InputHandoffs>,

fn cross_join<Other>(self, other: Other) -> CrossJoinPullSurface<Self, Other>where
Self: Sized,
Other: PullSurface,
Self::ItemOut: 'static + Eq + Clone,
Other::ItemOut: 'static + Eq + Clone,
Self::InputHandoffs: Extend<Other::InputHandoffs>,
<Self::InputHandoffs as Extend<Other::InputHandoffs>>::Extended: PortList<RECV> + PortListSplit<RECV, Self::InputHandoffs, Suffix = Other::InputHandoffs>,

fn fold_epoch<Init, Func, Out>(
self,
init: Init,
func: Func
) -> FoldEpochPullSurface<Self, Init, Func>where
Self: Sized,
Init: FnMut(&Context) -> Out,
Func: FnMut(&Context, Out, Self::ItemOut) -> Out,

Folds (accumulates) the input such that only the accumulated value is output each epoch. This operator should come after a inter-stratum handoff to ensure correct scheduling.

First, creates a accumulation value using the init method. Then, for each item in the epoch, run func to update the accumulation value. Finally, at the end of the scheduling quantum, release the single accumulation value to the next operator.

Example

use hydroflow::builder::prelude::*;

let mut builder = HydroflowBuilder::new();
builder.add_subgraph(
    "My Subgraph",
    [100_usize, 500, 30, 120]
        .into_iter()
        .into_hydroflow()
        .fold_epoch(|_ctx| 0, |_ctx, val, next| std::cmp::max(val, next))
        .pull_to_push()
        .for_each(|val| assert_eq!(500, val))
);

builder.build().run_available();

fn pull_to_push(self) -> PivotPushSurface<Self>where
Self: Sized,