Trait hydroflow::builder::surface::PushSurface

pub trait PushSurface: BaseSurface {
    type Output<Next>
    where
        Next: PushSurfaceReversed<ItemIn = Self::ItemOut>;

    fn push_to<Next>(self, next: Next) -> Self::Output<Next>
    where
        Next: PushSurfaceReversed<ItemIn = Self::ItemOut>;

    fn tee<NextA, NextB>(
        self, 
        next_a: NextA, 
        next_b: NextB
    ) -> Self::Output<TeePushSurfaceReversed<NextA, NextB>>
    where
        Self: Sized,
        Self::ItemOut: Clone,
        NextA: PushSurfaceReversed<ItemIn = Self::ItemOut>,
        NextB: PushSurfaceReversed<ItemIn = Self::ItemOut>,
        NextA::OutputHandoffs: Extend<NextB::OutputHandoffs>,
        <NextA::OutputHandoffs as Extend<NextB::OutputHandoffs>>::Extended: PortList<SEND> + PortListSplit<SEND, NextA::OutputHandoffs, Suffix = NextB::OutputHandoffs>,
    { ... }
    fn for_each_with_context<Func>(
        self, 
        func: Func
    ) -> Self::Output<ForEachPushSurfaceReversed<Func, Self::ItemOut>>
    where
        Self: Sized,
        Func: FnMut(&Context, Self::ItemOut),
    { ... }
    fn for_each<Func>(
        self, 
        func: Func
    ) -> Self::Output<ForEachPushSurfaceReversed<ForEachNoCtxFunc<Self, Func>, Self::ItemOut>>
    where
        Self: Sized,
        Func: FnMut(Self::ItemOut),
    { ... }
    fn partition_with_context<Func, NextA, NextB>(
        self, 
        func: Func, 
        next_a: NextA, 
        next_b: NextB
    ) -> Self::Output<PartitionPushSurfaceReversed<NextA, NextB, Func>>
    where
        Self: Sized,
        Func: FnMut(&Context, &Self::ItemOut) -> bool,
        NextA: PushSurfaceReversed<ItemIn = Self::ItemOut>,
        NextB: PushSurfaceReversed<ItemIn = Self::ItemOut>,
        NextA::OutputHandoffs: Extend<NextB::OutputHandoffs>,
        <NextA::OutputHandoffs as Extend<NextB::OutputHandoffs>>::Extended: PortList<SEND> + PortListSplit<SEND, NextA::OutputHandoffs, Suffix = NextB::OutputHandoffs>,
    { ... }
    fn partition<Func, NextA, NextB>(
        self, 
        func: Func, 
        next_a: NextA, 
        next_b: NextB
    ) -> Self::Output<PartitionNoCtxOutput<Self, Func, NextA, NextB>>
    where
        Self: Sized,
        Func: Fn(&Self::ItemOut) -> bool,
        NextA: PushSurfaceReversed<ItemIn = Self::ItemOut>,
        NextB: PushSurfaceReversed<ItemIn = Self::ItemOut>,
        NextA::OutputHandoffs: Extend<NextB::OutputHandoffs>,
        <NextA::OutputHandoffs as Extend<NextB::OutputHandoffs>>::Extended: PortList<SEND> + PortListSplit<SEND, NextA::OutputHandoffs, Suffix = NextB::OutputHandoffs>,
    { ... }
}

Required Associated Types

type Output<Next>
where
Next: PushSurfaceReversed<ItemIn = Self::ItemOut>

This should usually be a type which impls PushSurfaceReversed, but it is not enforced since we also need to return a Pivot in the end.

Required Methods

fn push_to<Next>(self, next: Next) -> Self::Output<Next> where
Next: PushSurfaceReversed<ItemIn = Self::ItemOut>,

Provided Methods

fn tee<NextA, NextB>(
self,
next_a: NextA,
next_b: NextB
) -> Self::Output<TeePushSurfaceReversed<NextA, NextB>> where
Self: Sized,
Self::ItemOut: Clone,
NextA: PushSurfaceReversed<ItemIn = Self::ItemOut>,
NextB: PushSurfaceReversed<ItemIn = Self::ItemOut>,
NextA::OutputHandoffs: Extend<NextB::OutputHandoffs>,
<NextA::OutputHandoffs as Extend<NextB::OutputHandoffs>>::Extended: PortList<SEND> + PortListSplit<SEND, NextA::OutputHandoffs, Suffix = NextB::OutputHandoffs>,

To create a output tee, use HydroflowBuilder::start_tee().

fn for_each_with_context<Func>(
self,
func: Func
) -> Self::Output<ForEachPushSurfaceReversed<Func, Self::ItemOut>> where
Self: Sized,
Func: FnMut(&Context, Self::ItemOut),

fn for_each<Func>(
self,
func: Func
) -> Self::Output<ForEachPushSurfaceReversed<ForEachNoCtxFunc<Self, Func>, Self::ItemOut>> where
Self: Sized,
Func: FnMut(Self::ItemOut),

fn partition_with_context<Func, NextA, NextB>(
self,
func: Func,
next_a: NextA,
next_b: NextB
) -> Self::Output<PartitionPushSurfaceReversed<NextA, NextB, Func>> where
Self: Sized,
Func: FnMut(&Context, &Self::ItemOut) -> bool,
NextA: PushSurfaceReversed<ItemIn = Self::ItemOut>,
NextB: PushSurfaceReversed<ItemIn = Self::ItemOut>,
NextA::OutputHandoffs: Extend<NextB::OutputHandoffs>,
<NextA::OutputHandoffs as Extend<NextB::OutputHandoffs>>::Extended: PortList<SEND> + PortListSplit<SEND, NextA::OutputHandoffs, Suffix = NextB::OutputHandoffs>,

fn partition<Func, NextA, NextB>(
self,
func: Func,
next_a: NextA,
next_b: NextB
) -> Self::Output<PartitionNoCtxOutput<Self, Func, NextA, NextB>> where
Self: Sized,
Func: Fn(&Self::ItemOut) -> bool,
NextA: PushSurfaceReversed<ItemIn = Self::ItemOut>,
NextB: PushSurfaceReversed<ItemIn = Self::ItemOut>,
NextA::OutputHandoffs: Extend<NextB::OutputHandoffs>,
<NextA::OutputHandoffs as Extend<NextB::OutputHandoffs>>::Extended: PortList<SEND> + PortListSplit<SEND, NextA::OutputHandoffs, Suffix = NextB::OutputHandoffs>,