Trait hydroflow::lang::collections::Collection

pub trait Collection<K, V> {
    type Keys<'s>: Iterator<Item = &'s K>
       where K: 's,
             Self: 's;
    type Entries<'s>: Iterator<Item = (&'s K, &'s V)>
       where K: 's,
             V: 's,
             Self: 's;
    type IntoEntries: Iterator<Item = (K, V)>;

    // Required methods
    fn get(&self, key: &K) -> Option<&V>;
    fn get_mut(&mut self, key: &K) -> Option<&mut V>;
    fn len(&self) -> usize;
    fn keys(&self) -> Self::Keys<'_>;
    fn entries(&self) -> Self::Entries<'_>;
    fn into_entries(self) -> Self::IntoEntries;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

An abstract collection, analagous to Map<K, V> or Set<K, ()>.

Required Associated Types

type Keys<'s>: Iterator<Item = &'s K> where K: 's, Self: 's

Iterator type returned by Self::keys.

type Entries<'s>: Iterator<Item = (&'s K, &'s V)> where K: 's, V: 's, Self: 's

Iterator type returned by Self::entries.

type IntoEntries: Iterator<Item = (K, V)>

Iterator type returned by Self::into_entries.

Required Methods

fn get(&self, key: &K) -> Option<&V>

Get a reference to the value corresponding to the key.

fn get_mut(&mut self, key: &K) -> Option<&mut V>

Gets a mutable reference to the value corresponding to the key.

fn len(&self) -> usize

Length of the collection.

fn keys(&self) -> Self::Keys<'_>

Returns an iterator over the keys in this collection.

fn entries(&self) -> Self::Entries<'_>

Returns an iterator over the key-value pairs in this collection.

fn into_entries(self) -> Self::IntoEntries

Turns this collection into an iterator over its key-value pairs. Similar to std::iter::IntoIterator but also works with sets for (K, ()) pairs.

Provided Methods

fn is_empty(&self) -> bool

If the collection is empty.