Macro hydroflow::var_type

macro_rules! var_type {
    () => { ... };
    ($a:ty $(, $b:ty)* $(,)?) => { ... };
}

Macro creating a variadic tuple type from a list of types.

var_expr! can be used to define simple types but will result in confusing errors for more complex types. Use this macro, var_type! instead.

Examples

use variadics::{var_expr, var_type};

// A simple variadic type. Although `var_expr!` would work in this case, it cannot handle
// more complex types i.e. ones with generics.
let list: var_type!(i32, bool, String) = Default::default();

// A more complex type:
let list: var_type!(
    &'static str,
    HashMap<i32, i32>,
    <std::vec::Vec<bool> as IntoIterator>::Item,
) = var_expr!("foo", HashMap::new(), false);

Unfortunately, expressions and types cannot be handled using the same macro due to the undefeated bastion of the turbofish.