Module hydroflow::scheduled

Hydroflow’s outer scheduled layer. Deals with inter-subgraph runtime data-passing and scheduling.

The most important item is the Hydroflow struct. Most of the items in this module are supporting the implementation of the Hydroflow struct and its operation.

Modules

context

Module for the user-facing Context object.

graph

Module for the Hydroflow struct and helper items.

graph_ext

Helper extensions for Hydroflow.

handoff

Module for all Handoff-related items.

input

net

This module contiains networking code.

port

Organizational module for Hydroflow Send/RecvCtx structs and Input/OutputPort structs.

query

Old query API, where each operator is a single subgraph. Deprecated.

reactor

Module for Reactor.

state

Module for StateHandle, part of the “state API”.

ticks

This module contains types to work with ticks.

Structs

HandoffId

A handoff’s ID. Invalid if used in a different graph::Hydroflow instance than the original that created it.

StateId

A staten handle’s ID. Invalid if used in a different graph::Hydroflow instance than the original that created it.

SubgraphId

A subgraph’s ID. Invalid if used in a different graph::Hydroflow instance than the original that created it.