Module hydroflow::scheduled::ctx

Organizational module for Hydroflow Send/RecvCtx structs and Input/OutputPort structs.

Structs

InputPort

Handle corresponding to a RecvCtx. Consumed by crate::scheduled::graph::Hydroflow::add_edge to construct the Hydroflow graph.

OutputPort

Handle corresponding to a SendCtx. Consumed by crate::scheduled::graph::Hydroflow::add_edge to construct the Hydroflow graph.

RecvCtx

Context provided to a compiled component for reading from an InputPort.

SendCtx

Context provided to a compiled component for writing to an OutputPort.