Struct hydroflow::scheduled::graph::Hydroflow

pub struct Hydroflow { /* fields omitted */ }

A Hydroflow graph. Owns, schedules, and runs the compiled subgraphs.

Implementations

impl Hydroflow

pub fn new() -> Self

Create an new empty Dataflow graph.

pub fn reactor(&self) -> Reactor

Returns a reactor for externally scheduling subgraphs, possibly from another thread.

pub fn tick(&mut self)

Runs the dataflow until no more work is currently available.

pub fn run(&mut self) -> Result<!, RecvError>

Run the dataflow graph, blocking until completion.

pub async fn run_async(&mut self) -> Result<!, RecvError>

Run the dataflow graph to completion asynchronously.

pub fn poll_events(&mut self) -> Result<(), RecvError>

Block and wait for an external event.

pub fn wake_all(&mut self)

TODO(mingwei): Hack to re-enqueue all subgraphs.

pub fn add_subgraph<F, R, W>(
&mut self,
subgraph: F
) -> (R::InputPort, W::OutputPort) where
F: 'static + FnMut(&Context<'_>, R::RecvCtx, W::SendCtx),
R: 'static + HandoffList,
W: 'static + HandoffList,

Adds a new compiled subgraph with the specified inputs and outputs.

See HandoffList for how to specify inputs and outputs.

pub fn add_n_in_m_out<F, R, W>(
&mut self,
n: usize,
m: usize,
f: F
) -> (Vec<InputPort<R>>, Vec<OutputPort<W>>) where
F: 'static + FnMut(&[&RecvCtx<R>], &[&SendCtx<W>]),
R: 'static + Handoff,
W: 'static + Handoff,

Adds a new compiled subraph with a variable number of inputs and outputs.

pub fn add_edge<H>(
&mut self,
output_port: OutputPort<H>,
input_port: InputPort<H>
) where
H: 'static + Handoff,

pub fn add_handoff<H: Handoff>(
&mut self,
pred_id: SubgraphId,
succ_id: SubgraphId
) -> HandoffId

pub fn add_state<T>(&mut self, state: T) -> StateHandle<T> where
T: Any,

impl Hydroflow

pub fn add_write_tcp_stream(
&mut self,
stream: TcpStream
) -> InputPort<VecHandoff<Message>>

pub fn add_read_tcp_stream(
&mut self,
stream: TcpStream
) -> OutputPort<VecHandoff<Message>>

pub fn add_tcp_stream(
&mut self,
stream: TcpStream
) -> (InputPort<VecHandoff<Message>>, OutputPort<VecHandoff<Message>>)