Trait hydroflow::scheduled::graph_ext::GraphExt

pub trait GraphExt {
    fn add_inout<F, R, W>(
        &mut self, 
        subgraph: F
    ) -> (InputPort<R>, OutputPort<W>)
    where
        F: 'static + FnMut(&Context<'_>, &RecvCtx<R>, &SendCtx<W>),
        R: 'static + Handoff,
        W: 'static + Handoff;

    fn add_binary_out<F, R, W1, W2>(
        &mut self, 
        subgraph: F
    ) -> (InputPort<R>, OutputPort<W1>, OutputPort<W2>)
    where
        F: 'static + FnMut(&Context<'_>, &RecvCtx<R>, &SendCtx<W1>, &SendCtx<W2>),
        R: 'static + Handoff,
        W1: 'static + Handoff,
        W2: 'static + Handoff;

    fn add_binary_in_binary_out<F, R1, R2, W1, W2>(
        &mut self, 
        subgraph: F
    ) -> (InputPort<R1>, InputPort<R2>, OutputPort<W1>, OutputPort<W2>)
    where
        F: 'static + FnMut(&Context<'_>, &RecvCtx<R1>, &RecvCtx<R2>, &SendCtx<W1>, &SendCtx<W2>),
        R1: 'static + Handoff,
        R2: 'static + Handoff,
        W1: 'static + Handoff,
        W2: 'static + Handoff;

    fn add_binary<F, R1, R2, W>(
        &mut self, 
        subgraph: F
    ) -> (InputPort<R1>, InputPort<R2>, OutputPort<W>)
    where
        F: 'static + FnMut(&Context<'_>, &RecvCtx<R1>, &RecvCtx<R2>, &SendCtx<W>),
        R1: 'static + Handoff,
        R2: 'static + Handoff,
        W: 'static + Handoff;

    fn add_binary_sink<F, R1, R2>(
        &mut self, 
        subgraph: F
    ) -> (InputPort<R1>, InputPort<R2>)
    where
        F: 'static + FnMut(&Context<'_>, &RecvCtx<R1>, &RecvCtx<R2>),
        R1: 'static + Handoff,
        R2: 'static + Handoff;

    fn add_sink<F, R>(&mut self, subgraph: F) -> InputPort<R>
    where
        F: 'static + FnMut(&Context<'_>, &RecvCtx<R>),
        R: 'static + Handoff;

    fn add_source<F, W>(&mut self, subgraph: F) -> OutputPort<W>
    where
        F: 'static + FnMut(&Context<'_>, &SendCtx<W>),
        W: 'static + Handoff;

    fn add_channel_input<T, W>(
        &mut self
    ) -> (Input<T, SyncSender<T>>, OutputPort<W>)
    where
        T: 'static,
        W: 'static + Handoff + CanReceive<T>;

    fn add_input<T, W>(&mut self) -> (Input<T, Buffer<T>>, OutputPort<W>)
    where
        T: 'static,
        W: 'static + Handoff + CanReceive<T>;

    fn add_input_from_stream<T, W, S>(&mut self, s: S) -> OutputPort<W>
    where
        S: 'static + Stream<Item = T> + Unpin,
        W: 'static + Handoff + CanReceive<T>;
}

Required methods

fn add_inout<F, R, W>(&mut self, subgraph: F) -> (InputPort<R>, OutputPort<W>) where
F: 'static + FnMut(&Context<'_>, &RecvCtx<R>, &SendCtx<W>),
R: 'static + Handoff,
W: 'static + Handoff,

Adds a new compiled subgraph with a single input and output, and returns the input/output handles.

fn add_binary_out<F, R, W1, W2>(
&mut self,
subgraph: F
) -> (InputPort<R>, OutputPort<W1>, OutputPort<W2>) where
F: 'static + FnMut(&Context<'_>, &RecvCtx<R>, &SendCtx<W1>, &SendCtx<W2>),
R: 'static + Handoff,
W1: 'static + Handoff,
W2: 'static + Handoff,

Adds a new compiled subgraph with one input and two outputs, and returns the input/output handles.

fn add_binary_in_binary_out<F, R1, R2, W1, W2>(
&mut self,
subgraph: F
) -> (InputPort<R1>, InputPort<R2>, OutputPort<W1>, OutputPort<W2>) where
F: 'static + FnMut(&Context<'_>, &RecvCtx<R1>, &RecvCtx<R2>, &SendCtx<W1>, &SendCtx<W2>),
R1: 'static + Handoff,
R2: 'static + Handoff,
W1: 'static + Handoff,
W2: 'static + Handoff,

Adds a new compiled subgraph with one input and two outputs, and returns the input/output handles.

fn add_binary<F, R1, R2, W>(
&mut self,
subgraph: F
) -> (InputPort<R1>, InputPort<R2>, OutputPort<W>) where
F: 'static + FnMut(&Context<'_>, &RecvCtx<R1>, &RecvCtx<R2>, &SendCtx<W>),
R1: 'static + Handoff,
R2: 'static + Handoff,
W: 'static + Handoff,

Adds a new compiled subraph with two inputs and a single output, and returns the input/output handles.

fn add_binary_sink<F, R1, R2>(
&mut self,
subgraph: F
) -> (InputPort<R1>, InputPort<R2>) where
F: 'static + FnMut(&Context<'_>, &RecvCtx<R1>, &RecvCtx<R2>),
R1: 'static + Handoff,
R2: 'static + Handoff,

Adds a new compiled subraph with two inputs and no outputs, and returns the input handles.

fn add_sink<F, R>(&mut self, subgraph: F) -> InputPort<R> where
F: 'static + FnMut(&Context<'_>, &RecvCtx<R>),
R: 'static + Handoff,

Adds a new compiled subgraph with one input and no outputs.

fn add_source<F, W>(&mut self, subgraph: F) -> OutputPort<W> where
F: 'static + FnMut(&Context<'_>, &SendCtx<W>),
W: 'static + Handoff,

Adds a new compiled subgraph with one output and no inputs.

fn add_channel_input<T, W>(
&mut self
) -> (Input<T, SyncSender<T>>, OutputPort<W>) where
T: 'static,
W: 'static + Handoff + CanReceive<T>,

fn add_input<T, W>(&mut self) -> (Input<T, Buffer<T>>, OutputPort<W>) where
T: 'static,
W: 'static + Handoff + CanReceive<T>,

Adds an “input” operator, returning a handle to insert data into it.

fn add_input_from_stream<T, W, S>(&mut self, s: S) -> OutputPort<W> where
S: 'static + Stream<Item = T> + Unpin,
W: 'static + Handoff + CanReceive<T>,