Module hydroflow::scheduled::handoff

Module for all Handoff-related items.

Modules

handoff_list

Module for variadic handoff port lists, PortList.

Structs

Iter

Wrapper around IntoIterator to avoid trait impl conflicts.

TeeingHandoff

VecHandoff

A Vec-based FIFO handoff.

Traits

CanReceive

Trait representing somethign which we can give an item to.

Handoff

Trait for handoffs to implement.

HandoffMeta

A handle onto the metadata part of a Handoff, with no element type.

TryCanReceive

Trait representing something which we can attempt to give an item to.