Trait hydroflow::scheduled::handoff::handoff_list::HandoffList

pub trait HandoffList: Variadic + Sealed { }

A variadic list of Handoff types, represented using a lisp-style tuple structure.

This trait is sealed and not meant to be implemented or used directly. Instead tuple lists (which already implement this trait) should be used, for example:

type MyHandoffList = (VecHandoff<usize>, (VecHandoff<String>, (TeeingHandoff<u32>, ())));

The var_expr! macro simplifies usage of this kind:

type MyHandoffList = var_expr!(VecHandoff<usize>, VecHandoff<String>, TeeingHandoff<u32>);