Trait hydroflow::scheduled::handoff::HandoffList

pub trait HandoffList: TypeList + Sealed {
    type InputHid: TypeList;
    type InputPort: TypeList;
    type RecvCtx: TypeList;
    type OutputHid: TypeList;
    type OutputPort: TypeList;
    type SendCtx: TypeList;
    fn make_input(sg_id: SubgraphId) -> (Self::InputHid, Self::InputPort);

    fn make_recv<'a>(
        handoffs: &'a [HandoffData], 
        input_hids: &Self::InputHid
    ) -> Self::RecvCtx;

    fn make_output(sg_id: SubgraphId) -> (Self::OutputHid, Self::OutputPort);

    fn make_send<'a>(
        handoffs: &'a [HandoffData], 
        output_hids: &Self::OutputHid
    ) -> Self::SendCtx;
}

A variadic list of Handoff types, represented using a lisp-style tuple structure.

This trait is sealed and not meant to be implemented or used directly. Instead tuple lists (which already implement this trait) should be used, for example:

type MyHandoffList = (VecHandoff<usize>, (VecHandoff<String>, (TeeingHandoff<u32>, ())));

The tl! (tuple list) macro simplifies usage of this kind:

type MyHandoffList = tl!(VecHandoff<usize>, VecHandoff<String>, TeeingHandoff<u32>);

Associated Types

type InputHid: TypeList

type InputPort: TypeList

type RecvCtx: TypeList

type OutputHid: TypeList

type OutputPort: TypeList

type SendCtx: TypeList

Required methods

fn make_input(sg_id: SubgraphId) -> (Self::InputHid, Self::InputPort)

fn make_recv<'a>(
handoffs: &'a [HandoffData],
input_hids: &Self::InputHid
) -> Self::RecvCtx

fn make_output(sg_id: SubgraphId) -> (Self::OutputHid, Self::OutputPort)

fn make_send<'a>(
handoffs: &'a [HandoffData],
output_hids: &Self::OutputHid
) -> Self::SendCtx