Module hydroflow::scheduled::port

Organizational module for Hydroflow Send/RecvCtx structs and Input/OutputPort structs.

Structs

Port

PortCtx

Enums

RECV

An uninstantiable type used to tag port Polarity as receive.

SEND

An uninstantiable type used to tag port Polarity as send.

Traits

Polarity

An empty trait used to denote Polarity: either send or receive.

Type Definitions

RecvCtx

RecvPort

SendCtx

SendPort