Module hydroflow::scheduled::state

Module for StateHandle, part of the “state API”.

Structs

StateHandle

A handle into a particular Hydroflow instance, referring to data inserted by add_state.

StateHandleErased

A state handle with the generic type parameter erased, allowing it to be stored in omogenous data structures. The type is tracked internally as data via TypeId.