Struct hydroflow::scheduled::ticks::TickDuration

pub struct TickDuration {
    pub ticks: i64,
}

The duration between two ticks.

TickDuration instances can be negative to allow for calculation of TickInstant instances in the past.

assert_eq!(TickInstant(1) + TickDuration::new(-1), TickInstant(0))

TickDuration instances can be added/subtracted to/from other TickDuration instances

assert_eq!(TickDuration::ZERO + TickDuration::ZERO, TickDuration::ZERO);
assert_eq!(
    TickDuration::ZERO + TickDuration::SINGLE_TICK,
    TickDuration::SINGLE_TICK
);
assert_eq!(
    TickDuration::SINGLE_TICK - TickDuration::ZERO,
    TickDuration::SINGLE_TICK
);
assert_eq!(
    TickDuration::SINGLE_TICK - TickDuration::SINGLE_TICK,
    TickDuration::ZERO
);
assert_eq!(
    TickDuration::ZERO - TickDuration::SINGLE_TICK,
    -TickDuration::SINGLE_TICK
);

Fields
ticks: i64

The length of the duration, measured in ticks.

Implementations

impl TickDuration

pub const ZERO: Self = _

A zero duration

It is the identity element for addition for both TickDuration and TickInstant (i.e. adding zero duration to a TickInstant or TickDuration results in the same TickInstant or TickDuration.

let ticks = TickInstant::new(100);
assert_eq!(ticks + TickDuration::ZERO, ticks);
assert_eq!(ticks - TickDuration::ZERO, ticks);

let duration = TickDuration::new(100);
assert_eq!(duration + TickDuration::ZERO, duration);
assert_eq!(duration - TickDuration::ZERO, duration);

pub const SINGLE_TICK: Self = _

A single tick duration.

It is the duration between two consecutive TickInstant instances.

assert_eq!(TickInstant(0) + TickDuration::SINGLE_TICK, TickInstant(1))

pub fn new(ticks: i64) -> TickDuration

Create a new TickDuration for the specified tick interval.

A negative duration allows for calculating TickInstants in the past and represents a backward movement in time.