Trait hydroflow::scheduled::HandoffList

pub trait HandoffList: Sealed {
    type RecvCtx;
    type InputPort;
    type Meta;
    type SendCtx;
    type OutputPort;
    fn make_input(op_id: OpId) -> (Self::RecvCtx, Self::InputPort);

    fn make_output() -> (Self::SendCtx, Self::OutputPort);
}

A variadic list of Handoff types, represented using a lisp-style tuple structure.

This trait is sealed and not meant to be implemented or used directly. Instead tuple lists (which already implement this trait) should be used, for example:

type MyHandoffList = (VecHandoff<usize>, (VecHandoff<String>, (NullHandoff, ())));

The tl! (tuple list) macro simplifies usage of this kind:

type MyHandoffList = tl!(VecHandoff<usize>, VecHandoff<String>, NullHandoff);

Associated Types

type RecvCtx

type InputPort

type Meta

type SendCtx

type OutputPort

Required methods

fn make_input(op_id: OpId) -> (Self::RecvCtx, Self::InputPort)

fn make_output() -> (Self::SendCtx, Self::OutputPort)