Module hydroflow::util

Helper utilities for the Hydroflow surface syntax.

Modules

connection

Functions

bind_udp_bytes

bind_udp_lines

collect_ready

Collects the immediately available items from the Stream into a FromIterator collection.

collect_ready_async

Collects the immediately available items from the Stream into a collection (Default + Extend).

deserialize_from_bytes

ipv4_resolve

ready_iter

serialize_to_bytes

sort_unstable_by_key_hrtb

tcp_bytes

Helper creates a TCP Stream and Sink for Bytes strings where each string is length-delimited.

tcp_bytestream

Helper creates a TCP Stream and Sink for undelimited streams of Bytes.

tcp_framed

Helper creates a TCP Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs.

tcp_lines

Helper creates a TCP Stream and Sink for strings delimited by newlines.

udp_bytes

Helper creates a UDP Stream and Sink for Bytes strings where each string is length-delimited.

udp_bytestream

Helper creates a UDP Stream and Sink for undelimited streams of Bytes.

udp_framed

Helper creates a UDP Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs. Also returns the bound UdpSocket, which will be different than the input UdpSocket if the input socket was set to port 0.

udp_lines

Helper creates a UDP Stream and Sink for strings delimited by newlines.

unbounded_channel

unix_bytes

Helper creates a Unix Stream and Sink for Bytes strings where each string is length-delimited.

unix_bytestream

Helper creates a Unix Stream and Sink for undelimited streams of Bytes.

unix_framed

Helper creates a Unix Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs.

unix_lines

Helper creates a Unix Stream and Sink for strings delimited by newlines.

Type Definitions

UdpBytesSink

UdpBytesStream

UdpFramedSink

UdpFramedStream

UdpLinesSink

UdpLinesStream

UdpSink

UdpStream