Trait hydroflow::util::simulation::MessageSender

pub trait MessageSender {
    // Required method
    fn send(&self, message: MessageWithAddress);
}

A message sender is used to send messages to an inbox on a host.

Required Methods

fn send(&self, message: MessageWithAddress)

Send a message to the inbox on the host.