Crate hydroflow

Re-exports

pub use bincode;

pub use bytes;

pub use futures;

pub use pusherator;

pub use rustc_hash;

pub use serde;

pub use serde_json;

pub use static_assertions;

pub use tokio;

pub use tokio_stream;

pub use tokio_util;

pub use variadics;

Modules

compiled

lang

props

scheduled

util

Helper utilities for the Hydroflow surface syntax.

Macros

assert_graphvis_snapshots

assert_var_impl

Asserts that the variable’s type implements the given traits.

datalog

hydroflow_parser

hydroflow_syntax

hydroflow_syntax_noemit

rassert

rassert_eq

var_args

Macro for pattern-matching with variadic tuples. This is mainly used for function arguments, but it can also be used in match, if let ..., and let ... else expressions.

var_expr

Macro creating a variadic tuple value from a list of expressions.

var_type

Macro creating a variadic tuple type from a list of types.