Trait hydroflow::builder::surface::PushSurfaceReversed

pub trait PushSurfaceReversed {
    type ItemIn;
    type OutputHandoffs: PortList<SEND>;
    type Build: PushBuild<OutputHandoffs = Self::OutputHandoffs, ItemIn = Self::ItemIn>;

    fn make_parts(self, ctx: &mut Context) -> (Self::OutputHandoffs, Self::Build);
}

This extra layer is needed due to the ownership order. In the functional chaining syntax each operator owns the previous (can only go in order things are called/defined), but in the end we need each pusherator to own the next pusherator which it’s pushing to.

This is the already-reversed, PushSurface does the actual reversing.

Required Associated Types

type ItemIn

type OutputHandoffs: PortList<SEND>

type Build: PushBuild<OutputHandoffs = Self::OutputHandoffs, ItemIn = Self::ItemIn>

Required Methods

fn make_parts(self, ctx: &mut Context) -> (Self::OutputHandoffs, Self::Build)