Module hydroflow::scheduled

Modules

collections

handoff

query

util

Structs

Hydroflow

A Hydroflow graph. Owns, schedules, and runs the compiled subgraphs.

InputPort

Handle corresponding to a RecvCtx. Consumed by Hydroflow::add_edge to construct the Hydroflow graph.

OutputPort

Handle corresponding to a SendCtx. Consumed by Hydroflow::add_edge to construct the Hydroflow graph.

RecvCtx

Context provided to a compiled component for reading from an InputPort.

SendCtx

Context provided to a compiled component for writing to an OutputPort.

Traits

HandoffList

A variadic list of Handoff types, represented using a lisp-style tuple structure.

Type Definitions

OpId