Struct hydroflow::scheduled::graph::Hydroflow

pub struct Hydroflow { /* private fields */ }

A Hydroflow graph. Owns, schedules, and runs the compiled subgraphs.

Implementations

impl Hydroflow

pub fn generate_mermaid(&self) -> String

pub fn write_mermaid(&self, write: &mut impl Write) -> Result

pub fn generate_dot(&self) -> String

pub fn write_dot(&self, w: &mut impl Write) -> Result

pub fn generate_json(&self) -> String

pub fn write_json(&self, write: &mut impl Write) -> Result

pub fn global_graph(&self) -> FlowPartitionedGraph

impl Hydroflow

pub fn new() -> Self

Create a new empty Hydroflow graph.

pub fn new_with_graph(serde_graph: &'static str) -> Self

Create a new empty Hydroflow graph with the given serde_graph JSON string.

pub fn serde_graph(&self) -> Option<&SerdeGraph>

pub fn reactor(&self) -> Reactor

Returns a reactor for externally scheduling subgraphs, possibly from another thread.

pub fn current_epoch(&self) -> usize

pub fn current_stratum(&self) -> usize

pub fn run_epoch(&mut self)

Runs the dataflow until the next epoch begins.

pub fn run_available(&mut self)

Runs the dataflow until no more work is immediately available. If the dataflow contains loops this method may run forever.

pub fn run_stratum(&mut self)

Runs the current stratum of the dataflow until no more work is immediately available.

pub fn next_stratum(&mut self) -> bool

Go to the next stratum which has work available, possibly the current stratum. Return true if more work is available, otherwise false if no work is immediately available on any strata.

pub fn run(&mut self) -> Option<!>

Runs the dataflow graph forever.

TODO(mingwei): Currently blockes forever, no notion of “completion.”

pub async fn run_async(&mut self) -> Option<!>

Runs the dataflow graph forever.

TODO(mingwei): Currently blockes forever, no notion of “completion.”

pub fn try_recv_events(&mut self) -> usize

Enqueues subgraphs triggered by external events without blocking.

Returns the number of subgraphs enqueued.

pub fn recv_events(&mut self) -> Option<NonZeroUsize>

Enqueues subgraphs triggered by external events, blocking until at least one subgraph is scheduled.

pub async fn recv_events_async(&mut self) -> Option<NonZeroUsize>

Enqueues subgraphs triggered by external events asynchronously, waiting until at least one subgraph is scheduled.

pub fn add_subgraph<Name, R, W, F>(
&mut self,
name: Name,
recv_ports: R,
send_ports: W,
subgraph: F
) -> SubgraphIdwhere
Name: Into<Cow<'static, str>>,
R: 'static + PortList<RECV>,
W: 'static + PortList<SEND>,
F: 'static + for<'ctx> FnMut(&'ctx Context, R::Ctx<'ctx>, W::Ctx<'ctx>),

pub fn add_subgraph_stratified<Name, R, W, F>(
&mut self,
name: Name,
stratum: usize,
recv_ports: R,
send_ports: W,
subgraph: F
) -> SubgraphIdwhere
Name: Into<Cow<'static, str>>,
R: 'static + PortList<RECV>,
W: 'static + PortList<SEND>,
F: 'static + for<'ctx> FnMut(&'ctx Context, R::Ctx<'ctx>, W::Ctx<'ctx>),

Adds a new compiled subgraph with the specified inputs and outputs.

TODO(mingwei): add example in doc.

pub fn add_subgraph_n_m<Name, R, W, F>(
&mut self,
name: Name,
recv_ports: Vec<RecvPort<R>>,
send_ports: Vec<SendPort<W>>,
subgraph: F
) -> SubgraphIdwhere
Name: Into<Cow<'static, str>>,
R: 'static + Handoff,
W: 'static + Handoff,
F: 'static + for<'ctx> FnMut(&'ctx Context, &'ctx [&'ctx RecvCtx<R>], &'ctx [&'ctx SendCtx<W>]),

Adds a new compiled subgraph with a variable number of inputs and outputs of the same respective handoff types.

pub fn add_subgraph_stratified_n_m<Name, R, W, F>(
&mut self,
name: Name,
stratum: usize,
recv_ports: Vec<RecvPort<R>>,
send_ports: Vec<SendPort<W>>,
subgraph: F
) -> SubgraphIdwhere
Name: Into<Cow<'static, str>>,
R: 'static + Handoff,
W: 'static + Handoff,
F: 'static + for<'ctx> FnMut(&'ctx Context, &'ctx [&'ctx RecvCtx<R>], &'ctx [&'ctx SendCtx<W>]),

Adds a new compiled subgraph with a variable number of inputs and outputs of the same respective handoff types.

pub fn make_edge<Name, H>(&mut self, name: Name) -> (SendPort<H>, RecvPort<H>)where
Name: Into<Cow<'static, str>>,
H: 'static + Handoff,

Creates a handoff edge and returns the corresponding send and receive ports.

pub fn add_state<T>(&mut self, state: T) -> StateHandle<T>where
T: Any,

pub fn context_mut(&mut self, sg_id: SubgraphId) -> &mut Context

Gets a exclusive (mut) ref to the internal context, setting the subgraph ID.

pub fn add_dependencies(&mut self, sg_id: SubgraphId, deps: FlowGraph)

impl Hydroflow

pub fn spawn_task<Fut>(&mut self, future: Fut) -> Result<(), TryCurrentError>where
Fut: Future<Output = ()> + Send + 'static,

pub fn abort_tasks(&mut self)

pub fn join_tasks(&mut self) -> impl '_ + Future

impl Hydroflow

pub async fn inbound_tcp_vertex_port<T>(
&mut self,
port: u16
) -> RecvPort<VecHandoff<T>>where
T: 'static + DeserializeOwned + Send,

pub async fn inbound_tcp_vertex<T>(&mut self) -> (u16, RecvPort<VecHandoff<T>>)where
T: 'static + DeserializeOwned + Send,

pub async fn outbound_tcp_vertex<T>(
&mut self
) -> SendPort<VecHandoff<(Address, T)>>where
T: 'static + Serialize + Send,

impl Hydroflow

pub fn add_write_tcp_stream(
&mut self,
stream: TcpStream
) -> SendPort<VecHandoff<Message>>

pub fn add_read_tcp_stream(
&mut self,
stream: TcpStream
) -> RecvPort<VecHandoff<Message>>

pub fn add_tcp_stream(
&mut self,
stream: TcpStream
) -> (SendPort<VecHandoff<Message>>, RecvPort<VecHandoff<Message>>)