Type Alias hydroflow::scheduled::port::RecvPort

pub type RecvPort<H> = Port<RECV, H>;

Recv-specific variant of Port. An input port.

Aliased Type

struct RecvPort<H> { /* private fields */ }

Implementations

impl<T: Clone> RecvPort<TeeingHandoff<T>>

Methods for TeeingHandoff teeing and dropping.

pub fn tee(&self, hf: &mut Hydroflow<'_>) -> RecvPort<TeeingHandoff<T>>

Tees this TeeingHandoff, given the Hydroflow instance it belongs to.

pub fn drop(self, hf: &mut Hydroflow<'_>)

Marks this output of a TeeingHandoff as dropped so that no more data will be sent to it, given the Hydroflow instance it belongs to.

It is recommended to not not use this method and instead simply avoid teeing a TeeingHandoff when it is not needed.