Module hydroflow::util

Helper utilities for the Hydroflow surface syntax.

Modules

unsync

Unsync (i.e. single-threaded) synchronization primitives.

Functions

bind_udp_bytes

Returns a length-delimited bytes Sink, Stream, and SocketAddr bound to the given address. The input addr may have a port of 0, the returned SocketAddr will have the chosen port.

bind_udp_lines

Returns a newline-delimited bytes Sink, Stream, and SocketAddr bound to the given address. The input addr may have a port of 0, the returned SocketAddr will have the chosen port.

collect_ready

Collects the immediately available items from the Stream into a FromIterator collection.

collect_ready_async

Collects the immediately available items from the Stream into a collection (Default + Extend).

deserialize_from_bytes

Serialize a message from bytes using bincode.

ipv4_resolve

Resolve the ipv4 SocketAddr from an IP or hostname string.

ready_iter

Returns an Iterator of any immediately available items from the [Stream].

serialize_to_bytes

Serialize a message to bytes using bincode.

sort_unstable_by_key_hrtb

Sort a slice using a key fn which returns references.

tcp_bytes

Helper creates a TCP Stream and Sink for Bytes strings where each string is length-delimited.

tcp_bytestream

Helper creates a TCP Stream and Sink for undelimited streams of Bytes.

tcp_framed

Helper creates a TCP Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs.

tcp_lines

Helper creates a TCP Stream and Sink for strings delimited by newlines.

udp_bytes

Helper creates a UDP Stream and Sink for Bytes strings where each string is length-delimited.

udp_bytestream

Helper creates a UDP Stream and Sink for undelimited streams of Bytes.

udp_framed

Helper creates a UDP Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs. Also returns the bound UdpSocket, which will be different than the input UdpSocket if the input socket was set to port 0.

udp_lines

Helper creates a UDP Stream and Sink for Strings delimited by newlines.

unbounded_channel

Returns a channel as a (1) unbounded sender and (2) unbounded receiver Stream for use in Hydroflow.

unix_bytes

Helper creates a Unix Stream and Sink for Bytes strings where each string is length-delimited.

unix_bytestream

Helper creates a Unix Stream and Sink for undelimited streams of Bytes.

unix_framed

Helper creates a Unix Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs.

unix_lines

Helper creates a Unix Stream and Sink for strings delimited by newlines.

unsync_channel

Returns an unsync channel as a (1) sender and (2) receiver Stream for use in Hydroflow.

Type Definitions

UdpBytesSink

A UDP undelimited bytes Sink (sending).

UdpBytesStream

A UDP undelimited bytes Stream (receiving).

UdpFramedSink

A framed UDP Sink (sending).

UdpFramedStream

A framed UDP Stream (receiving).

UdpLinesSink

A UDP newline-delimited String Sink (sending).

UdpLinesStream

A UDP newline-delimited String Stream (receivng).

UdpSink

A UDP length-delimited frame Sink (sending).

UdpStream

A UDP length-delimited frame Stream (receiving).