Module hydroflow::util

Helper utilities for the Hydroflow surface syntax.

Modules

clear

Module for the Clear trait.

cli

demux_enum

Traits for the demux_enum derive and operator.

monotonic_map

Module for MonotonicMap.

multiset

A multiset backed by a HashMap

sparse_vec

A vector that supports efficient deletion without reordering all subsequent items.

unsync

Unsync (i.e. single-threaded) synchronization primitives.

Structs

DroppableChild

When a Child is dropped normally nothing happens but in unit tests you usually want to terminate the child and wait for it to terminate. DroppableChild does that for us.

MonotonicFn

A wrapper christening a closure as a monotonic function

Morphism

A wrapper christening a closure as a lattice morphism

Enums

Persistence

Persit or delete tuples

PersistenceKeyed

Persit or delete key-value pairs

Functions

bind_tcp

Create a listening tcp socket, and then as new connections come in, receive their data and forward it to a queue.

bind_tcp_bytes

Returns a newline-delimited bytes Sender, Receiver, and SocketAddr bound to the given address. The input addr may have a port of 0, the returned SocketAddr will be the address of the newly bound endpoint. The inbound connections can be used in full duplex mode. When a (T, SocketAddr) pair is fed to the Sender returned by this function, the SocketAddr will be looked up against the currently existing connections. If a match is found then the data will be sent on that connection. If no match is found then the data is silently dropped.

bind_tcp_lines

This is the same thing as bind_tcp_bytes except instead of using a length-delimited encoding scheme it uses new lines to separate frames.

bind_udp_bytes

Returns a length-delimited bytes Sink, Stream, and SocketAddr bound to the given address. The input addr may have a port of 0, the returned SocketAddr will have the chosen port.

bind_udp_lines

Returns a newline-delimited bytes Sink, Stream, and SocketAddr bound to the given address. The input addr may have a port of 0, the returned SocketAddr will have the chosen port.

collect_ready

Collects the immediately available items from the Stream into a FromIterator collection.

collect_ready_async

Collects the immediately available items from the Stream into a collection (Default + Extend).

connect_tcp

This is the inverse of bind_tcp, when messages enqueued into the returned sender, tcp sockets will be created and connected as necessary to send out the requests. As the responses come back, they will be forwarded to the returned receiver.

connect_tcp_bytes

This is inverse of bind_tcp_bytes. (Bytes, SocketAddr) pairs fed to the returned Sender will initiate new tcp connections to the specified SocketAddr. These connections will be cached and reused, so that there will only be one connection per destination endpoint. When the endpoint sends data back it will be available via the returned Receiver

connect_tcp_lines

This is the same thing as connect_tcp_bytes except instead of using a length-delimited encoding scheme it uses new lines to separate frames.

deserialize_from_bytes

Serialize a message from bytes using bincode.

ipv4_resolve

Resolve the ipv4 SocketAddr from an IP or hostname string.

iter_batches_stream

Returns an [Stream] that emits n items at a time from iter at a time, yielding in-between. This is useful for breaking up a large iterator across several ticks: source_iter(...) always releases all items in the first tick. However using iter_batches_stream with source_stream(...) will cause n items to be released each tick. (Although more than that may be emitted if there are loops in the stratum).

ready_iter

Returns an Iterator of any immediately available items from the [Stream].

run_cargo_example

rust examples are meant to be run by people and have a natural interface for that. This makes unit testing them cumbersome. This function wraps calling cargo run and piping the stdin/stdout of the example to easy to handle returned objects. The function also returns a DroppableChild which will ensure that the child processes will be cleaned up appropriately.

serialize_to_bytes

Serialize a message to bytes using bincode.

sort_unstable_by_key_hrtb

Sort a slice using a key fn which returns references.

tcp_bytes

Helper creates a TCP Stream and Sink for Bytes strings where each string is length-delimited.

tcp_bytestream

Helper creates a TCP Stream and Sink for undelimited streams of Bytes.

tcp_framed

Helper creates a TCP Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs.

tcp_lines

Helper creates a TCP Stream and Sink for strings delimited by newlines.

udp_bytes

Helper creates a UDP Stream and Sink for Bytes strings where each string is length-delimited.

udp_bytestream

Helper creates a UDP Stream and Sink for undelimited streams of Bytes.

udp_framed

Helper creates a UDP Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs. Also returns the bound UdpSocket, which will be different than the input UdpSocket if the input socket was set to port 0.

udp_lines

Helper creates a UDP Stream and Sink for Strings delimited by newlines.

unbounded_channel

Returns a channel as a (1) unbounded sender and (2) unbounded receiver Stream for use in Hydroflow.

unix_bytes

Helper creates a Unix Stream and Sink for Bytes strings where each string is length-delimited.

unix_bytestream

Helper creates a Unix Stream and Sink for undelimited streams of Bytes.

unix_framed

Helper creates a Unix Stream and Sink from the given socket, using the given Codec to handle delineation between inputs/outputs.

unix_lines

Helper creates a Unix Stream and Sink for strings delimited by newlines.

unsync_channel

Returns an unsync channel as a (1) sender and (2) receiver Stream for use in Hydroflow.

wait_for_process_output

When a child process is spawned often you want to wait until the child process is ready before moving on. One way to do that synchronization is by waiting for the child process to output something and match regex against that output. For example, you could wait until the child process outputs “Client live!” which would indicate that it is ready to receive input now on stdin.

Type Aliases

TcpFramedSink

A framed TCP Sink (sending).

TcpFramedStream

A framed TCP Stream (receiving).

UdpBytesSink

A UDP undelimited bytes Sink (sending).

UdpBytesStream

A UDP undelimited bytes Stream (receiving).

UdpFramedSink

A framed UDP Sink (sending).

UdpFramedStream

A framed UDP Stream (receiving).

UdpLinesSink

A UDP newline-delimited String Sink (sending).

UdpLinesStream

A UDP newline-delimited String Stream (receivng).

UdpSink

A UDP length-delimited frame Sink (sending).

UdpStream

A UDP length-delimited frame Stream (receiving).